+++ Essentials/Pathfinding/Pathfinding/Core/Pathfinding.cs
@@ -476,18 +476,11 @@ namespace ME.ECS.Pathfinding {
             
             for (int i = 0; i < tasks.Length; ++i) {
 
-                var item = tasks[i];
-                if (item.burstEnabled == true) {
-                    
-                    job.Execute(i);
-                    item.isValid = false;
-                    tasks[i] = item;
+                job.Execute(i);
 
-                }
-                
             }
             
-            var jobHandle = job.Schedule(tasks.Length, 64);
+            /*var jobHandle = job.Schedule(tasks.Length, 64);
             jobHandle.Complete();
             
             for (int i = 0; i < tasks.Length; ++i) {
@@ -500,7 +493,7 @@ namespace ME.ECS.Pathfinding {
 
                 }
                 
-            }
+            }*/
 
             results = Pathfinding.results;