+++ Runtime/Core/Filters/Archetypes/FiltersArchetypeStorage.cs
@@ -1265,7 +1265,8 @@ namespace ME.ECS.FiltersArchetype {
                 var filterStaticData = world.GetFilterStaticData(filter.id);
                 if (filterStaticData.isCreated == false) continue;
                 
-                if (FiltersArchetypeStorage.IsEquals(filterStaticData.data.contains, filterBuilder.data.contains) == true &&
+                if (filterStaticData.data.withinTicks == filterBuilder.data.withinTicks &&
+                    FiltersArchetypeStorage.IsEquals(filterStaticData.data.contains, filterBuilder.data.contains) == true &&
                     FiltersArchetypeStorage.IsEquals(filterStaticData.data.notContains, filterBuilder.data.notContains) == true &&
                     FiltersArchetypeStorage.IsEquals(filterStaticData.data.notContainsShared, filterBuilder.data.notContainsShared) == true &&
                     FiltersArchetypeStorage.IsEquals(filterStaticData.data.containsShared, filterBuilder.data.containsShared) == true &&