+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Version 2.5
 * Essentials.GOAP: Job replaced with static burst method
-* Runtime.Core: CreateFromData now with lambda support
+* Runtime.Core: 
+  * Filter WithinTicks method added into builder
+  * CreateFromData now with lambda support
 * Runtime.Core: FilterBag GetIndexByEntityId methods added
 * Essentials.Pathfinding: NavMeshPath DataObject changed to ListCopyable
 * GOAPAction: fixed point math support added [Oleg-Grim]